% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajout_donnees_bio.R
\name{ajout_donnees_bio}
\alias{ajout_donnees_bio}
\title{Ajouter les données récentes aux données historiques de données biologiques.}
\usage{
ajout_donnees_bio(donnees_2020_et_moins = NULL, input_2021_et_plus_fichier)
}
\arguments{
\item{donnees_2020_et_moins}{\code{data.frame} contenant les données historique consolidées}

\item{input_2021_et_plus_fichier}{chemin du fichier contenant les données récentes}
}
\value{
le data.frame() consolidé
}
\description{
Les nouvelles données sont présentent dans le dossier "input_2021_et_plus_dir", chaque nouvelle années dans un dossier avec comme nom "Saison AAAA"
}
