% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_capture_totale.R
\name{captureTotale}
\alias{captureTotale}
\title{Estime la capture anuelle totale pour chacune des espèces sébastes, morue, ogac et turbot et produit un graphique de ces captures.}
\usage{
captureTotale(
  capt.pecheur,
  nbPecheursMoy.strate,
  nbJour.strate,
  legende = c(FALSE, TRUE, FALSE, FALSE),
  langue = c("fr", "en", "bil")
)
}
\arguments{
\item{capt.pecheur}{un data.frame() de pêcheurs avec leur capture par espece}

\item{legende}{dans quel(s) graphique doit apparaitre la légende}

\item{langue}{trois choix de langue du texte dans le graphique, soit français ('fr'), anglais('en') ou bilingue('bil')}

\item{nbPecheurMoy.strate}{moyenne du nombre de pêcheurs estimé par strate}

\item{nbJours.strate}{nombre de jour que dure la saison, par strate}
}
\value{
data.frame() des captures totales par strate
}
\description{
Pour chaque strate année/site/semaineFinSemaine, un nombre moyen de captures par pêcheur est calculé.
Ce nombre est multiplié par le nombre de moyen de pêcheurs pour la strate et par le nombre de jour calculé pour une saison complète de la strate.
}
