% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_dates_ech.R
\name{graph_dates_ech}
\alias{graph_dates_ech}
\title{Calcule la plage de dates où des échantillons sont récoltés pour chaque année et produit le graphique correspondant}
\usage{
graph_dates_ech(
  ech,
  visites,
  ouvertureOfficielle,
  langue = c("fr", "en", "bil")
)
}
\arguments{
\item{ech}{data.frame() contenant les données récoltées par les échantillonneurs}

\item{visites}{data.frame() contenant les données récoltées par les échantillonneurs, une ligne par visite}

\item{ouvertureOfficielle}{dataframe() contenant les années, mois et jours d'ouverture et
de fermeture de la pêche blanche pour chaque année de gestion}

\item{langue}{trois choix de langue du texte dans le graphique, soit français ('fr'), anglais('en') ou bilingue('bil')}
}
\value{
Ne retourne rien, produit un graphique
}
\description{
Calcule la plage de dates où des échantillons sont récoltés pour chaque année et produit le graphique correspondant
}
