% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effort_peche.R
\name{graph_effort_peche}
\alias{graph_effort_peche}
\title{Calcule et produit la fréquentation annuelle total de l'activité de pêche réacréative hivernale sur le Saguenay.}
\usage{
graph_effort_peche(nbPecheurs, nbJours, langue = c("fr", "en", "bil"))
}
\arguments{
\item{nbPecheurs}{data.frame() contenant le nombre moyen de pêcheurs par site, année et moment de la semaine}

\item{nbJours}{data.frame() détaillant le nombre de jours de pêche selon l'année et le moment de la semaine}

\item{langue}{trois choix de langue du texte dans le graphique, soit français ('fr'), anglais('en') ou bilingue('bil')}
}
\value{
data.frame() du nombre total de pêcheurs-jour par année, et pour le secteur de La Baie et le PMSSL
}
\description{
Calcule et produit la fréquentation annuelle total de l'activité de pêche réacréative hivernale sur le Saguenay.
}
