% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effort_peche_barplot.R
\name{graph_effort_peche_barplot}
\alias{graph_effort_peche_barplot}
\title{Calcule la fréquentation annuelle total}
\usage{
graph_effort_peche_barplot(nbPecheursTot, langue = c("fr", "en", "bil"))
}
\arguments{
\item{nbPecheursTot}{data.frame() contenant le nombre annuel total de pêcheurs-jour pour tour le saguenay et
par secteur (La Baie et PMSSL)}

\item{langue}{trois choix de langue du texte dans le graphique, soit français ('fr'), anglais('en') ou bilingue('bil')}
}
\value{
data.frame() du nombre total de pêcheurs-jour par année, et pour le secteur de La Baie et le PMSSL
}
\description{
Calcule la fréquentation annuelle total
}
