% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effort_peche_par_site.R
\name{graph_effort_peche_par_site}
\alias{graph_effort_peche_par_site}
\title{Calcule la fréquentation annuelle des différents sites de pêche, selon les vistes des échantillonneurs.}
\usage{
graph_effort_peche_par_site(ech, visites, langue = c("fr", "en", "bil"))
}
\arguments{
\item{ech}{data.frame() contenant les données récoltées par les échantillonneurs}

\item{visites}{data.frame() contenant les visites réalisées par les échantillonneurs}

\item{langue}{trois choix de langue du texte dans le graphique, soit français ('fr'), anglais('en') ou bilingue('bil')}
}
\value{
un array() de 4 dimension (site, année, sfs et statistique) fournissant les statistiques de moyenne, de sd et de n.
}
\description{
Les sites sont c("AnseBenjamin", "AnseStJean", "GrandeBaie","LesBattures", "RiviereEternite", "SteRose",
"StFelix", "StFulgence")
}
