% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_donnees_bio_pre2021.R
\name{lire_donnees_bio_pre2021}
\alias{lire_donnees_bio_pre2021}
\title{Lit les différents fichiers de données biologiques utilisés jusqu'en 2021 dans l'évaluation de stock,
standardise les noms de site et d'espèce, consolide le tout dans un data.frame().}
\usage{
lire_donnees_bio_pre2021(input_dir)
}
\arguments{
\item{input_dir}{Chemin vers le dossier où sont les fichiers de données brutes}
}
\value{
le data.frame() consolidé
}
\description{
Avant 2001, les morues franches et ogac sont confondus.
À partir de 2001, les échantillonneurs sont formés à faire la différence.
Aucune données en 2021 pour cause de pandémie.
Commentaires supplémentaires: - mauvaises dates en 2000
- les dates sont à valder 2002, 1 janier par défaut
- différents turbot sans date et site en 2007
- problèmes de dates en 2009, sort bien avec excel, mais pas ici
- deux formats de dates à gérer en 2010, 2015, 2016, 2017, 2018
- certaines date au début 2014 dans les fichiers de 2015 et 2016
- certaines date en 2108 dans le fichier 2018
}
