% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_pre2021.R
\name{lire_donnees_pre2021}
\alias{lire_donnees_pre2021}
\title{Créer les fichier .csv consolidant données historiques de chacune des trois sources (pré-2021)}
\usage{
lire_donnees_pre2021(dir_input, dir_output)
}
\arguments{
\item{dir_input}{Chemin vers le dossier contenant les fichiers de données brutes}

\item{dir_output}{Chemin vers le dossier où sauvegarder les fichier .csv produits}
}
\value{
Aucun objet n'est retourné.
Les fichiers générés sont sauvegardés dans le dossier spécifié par \code{output_dir}
}
\description{
Cette fonction lit les fichiers de données dans un dossier source,
applique les transformations nécessaires, puis sauvegarde en format .csv dans le dossier destination
}
\details{
Ce code n'a pas à être exécuté pour réaliser l'évaluation, mais permet le suivi du traitement des données anciennes.
Les trois sources de données consolidées ici sont:
\itemize{
\item Les données biologiques
\item Les journaux de bord
\item les échantillonneurs
}
}
\examples{
##À venir
}
