% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_echantillonneurs_pre2021.R
\name{lire_echantillonneurs_pre2021}
\alias{lire_echantillonneurs_pre2021}
\title{Lit les différents fichiers des échantillonneurs utilisés jusqu'en 2020 dans l'évaluation de stock,
standardise les noms de site, consolide le tout dans un data.frame().}
\usage{
lire_echantillonneurs_pre2021(input_dir)
}
\arguments{
\item{input_dir}{Chemin vers le dossier où sont les fichiers de données brutes}
}
\value{
une list() avec le data.frame() consolidé ($ech) et le data.frame() incluant des données mise de côté pour l'évaluation mais potentiellement
d'intérêt pour d'autres travaux
}
\description{
Aucune données en 2021 pour cause de pandémie.
}
