% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_glaces_du_fjord.R
\name{lire_glaces_du_fjord}
\alias{lire_glaces_du_fjord}
\title{Lire les données issues de l'application Glaces du Fjord.}
\usage{
lire_glaces_du_fjord(input_2021_et_plus_dir)
}
\arguments{
\item{input_2021_et_plus_dir}{chemin du dossier contenant des dossier "Saison 20xx" avec les nouvelles données}
}
\value{
le data.frame() consolidé
}
\description{
Les données sont produites par Contact Nature et sont obtenues sur demande à ceux-ci
Les données sont présentent dans le dossier "input_2021_et_plus_dir", chaque nouvelle années dans un dossier avec comme nom "Saison AAAA"
}
\details{
Note à développer: associer à des sites ou non?
}
