% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_journaux_de_bord_pre2021.R
\name{lire_journaux_de_bord_pre2021}
\alias{lire_journaux_de_bord_pre2021}
\title{Lit les différents fichiers de journaux de bord utilisés jusqu'en 2021 dans l'évaluation de stock,
standardise les noms de site et d'espèce, consolide le tout dans un data.frame().}
\usage{
lire_journaux_de_bord_pre2021(
  input_dir = file.path("S:", "Saguenay", "Pêche hivernale", "JB")
)
}
\arguments{
\item{input_dir}{Chemin vers le dossier où sont les fichiers de données brutes}
}
\value{
le data.frame() consolidé
}
\description{
Lit les différents fichiers de journaux de bord utilisés jusqu'en 2021 dans l'évaluation de stock,
standardise les noms de site et d'espèce, consolide le tout dans un data.frame().
}
