% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiser_nom_engin.R
\name{standardiser_nom_engin}
\alias{standardiser_nom_engin}
\title{Standardise l'orthographe des noms d'engins}
\usage{
standardiser_nom_engin(engins)
}
\arguments{
\item{engins}{vector() des engins à identifier et valider}
}
\value{
un vector() de noms standardisés
}
\description{
Standardise l'orthographe des noms d'engins
}
