% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiser_nom_espece.R
\name{standardiser_nom_espece}
\alias{standardiser_nom_espece}
\title{Standardise l'orthographe des noms d'espèce}
\usage{
standardiser_nom_espece(espece)
}
\arguments{
\item{espece}{vector() des espèces à identifier et valider}
}
\value{
un vector() de noms standardisés
}
\description{
Avant 2001, les morues franches et ogac sont confondus.
À partir de 2001, les échantillonneurs sont formés à faire la différence.
}
