% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiser_nom_site.R
\name{standardiser_nom_site}
\alias{standardiser_nom_site}
\title{Standardise l'orthographe des appellations de lieux.}
\usage{
standardiser_nom_site(sites)
}
\arguments{
\item{sites}{Vector() des sites à identifier et valider}
}
\value{
Une liste de 2 éléments, l'orthographe des noms originaux standardisés ($site.ori) et les noms regroupés en 8 villages de pêche ($site)
}
\description{
StFulgence = noSite(1)
AnseBenjamin = noSite(2)
GrandeBaie = noSite(3)
LesBattures = noSite(4)
SteRose = noSite(5)
StFelix = noSite(6)
RiviereEternite = noSite(7)
AnseStJean = noSite(8)
Indetermine= noSite(99)
}
