% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajout_echantillonneurs.R
\name{ajout_echantillonneurs}
\alias{ajout_echantillonneurs}
\title{Ajouter les données récentes aux données historiques des journaux de bord.}
\usage{
ajout_echantillonneurs(
  donnees_2020_et_moins = NULL,
  input_2021_et_plus_fichier
)
}
\arguments{
\item{donnees_2020_et_moins}{\code{data.frame} contenant les données historique consolidées}

\item{fichier_2021_et_plus_fichier}{chemin le fichier contenant les données récentes}
}
\value{
le data.frame() consolidé
}
\description{
Les nouvelles données sont présentent dans le fichier "input_2021_et_plus_fichier"
}
