% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajout_journaux_de_bord.R
\name{ajout_journaux_de_bord}
\alias{ajout_journaux_de_bord}
\title{Ajouter les données récentes aux données historiques de journaux de bords.}
\usage{
ajout_journaux_de_bord(donnees_2020_et_moins = NULL, input_2021_et_plus_dir)
}
\arguments{
\item{donnees_2020_et_moins}{\code{data.frame} contenant les données historique consolidées}

\item{input_2021_et_plus_dir}{chemin du dossier contenant des dossier "Saison 20xx" avec les nouvelles données}
}
\value{
le data.frame() consolidé
}
\description{
Les nouvelles données sont présentent dans le dossier "input_2021_et_plus_dir", chaque nouvelle années dans un dossier avec comme nom "Saison AAAA"
}
