% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_de_peche.R
\name{effort_de_peche}
\alias{effort_de_peche}
\title{Calcul l'effort de pêche sur le Saguenay et produit les graphiques correspondants.}
\usage{
effort_de_peche(ech)
}
\arguments{
\item{ech}{données récolétées dans le volet 1 du projet (échantillonneurs).}
}
\value{
Aucun objet est retourné mais deux fichiers sont créés (image et données).
}
\description{
Calcul l'effort de pêche sur le Saguenay et produit les graphiques correspondants.
}
\examples{
##À venir
}
