% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_donnees.R
\name{lire_donnees}
\alias{lire_donnees}
\title{Lire les données disponibles pour l'évaluation}
\usage{
lire_donnees(donnees_2020_et_moins_dir = NULL, dir_input, dir_output)
}
\arguments{
\item{donnees_2020_et_moins_dir}{chemin du répertoire des données récentes}

\item{dir_input}{chemin vers le dossier où sont situées les donnée}

\item{dir_output}{chemin vers le dossier où sont enregistrés les fichiers d'info consolidés}
}
\value{
une liste des données de chacune des sources
}
\description{
Quatres sources sont combinées dans la liste retournée, soit:
\itemize{
\item Les enquêtes des échantillonneurs
\item Les journaux de bord distribués à certains pêcheurs
\item Les données biologiques liées aux mesures de poissons individuels
\item Les informations collectées par l'application "Glaces du fjord"
}
}
\examples{
##À venir
}
